requirements.txt
<<<<<<< HEAD
blinker==1.9.0
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
flask-marshmallow==1.3.0
=======
aniso8601==10.0.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-RESTful==0.3.10
>>>>>>> 5b3cc0b (Added requirements.txt and .gitignore)
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
<<<<<<< HEAD
marshmallow==3.26.0
openpyxl==3.1.5
packaging==24.2
SQLAlchemy==2.0.37
=======
marshmallow==3.26.1
packaging==24.2
PyJWT==2.10.1
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
>>>>>>> 5b3cc0b (Added requirements.txt and .gitignore)
typing_extensions==4.12.2
Werkzeug==3.1.3
